requirements.txt
fastapi
uvicorn
openai-whisper
language-tool-python
python-multipart
